git-svn-id: http://svn.l2jfree.com/l2jfree@8768 6ff00ba9-25cd-4355-987e-099c62bc2628

+++ trunk/l2jfree-datapack/sql/auction_bid.sql
@@ -4,7 +4,7 @@ CREATE TABLE IF NOT EXISTS `auction_bid` (
   `bidderId` INT NOT NULL DEFAULT 0,
   `bidderName` VARCHAR(35) NOT NULL,
   `clan_name` VARCHAR(45) NOT NULL,
-  `maxBid` INT UNSIGNED NOT NULL DEFAULT 0,
+  `maxBid` BIGINT UNSIGNED NOT NULL DEFAULT 0,
   `time_bid` BIGINT NOT NULL DEFAULT 0,
   PRIMARY KEY (`auctionId`, `bidderId`),
   KEY `id` (`id`)
